+++ cmd/main.go
@@ -4,10 +4,15 @@
 package main
 
 import (
+	"errors"
 	"fmt"
+	"os"
 )
 
+var db idb
+
 func main() {
+
 	// Welcome.
 	fmt.Println()
 	fmt.Println("MIGRATABLE")
@@ -16,16 +21,36 @@ func main() {
 	// Gather the user request.
 	c, err := getConfig()
 	if err == nil {
-		// Action it.
 		c.describe()
+
+		// Fetch the migrations.
 		var m *migrations
 		m, err = loadMigrations(c.folder)
 		if err == nil {
 			fmt.Printf("  Migrations loaded: %v\n", len(*m))
+
+			// Get a database connection.
+			connStr := os.Getenv(c.connEnv)
+			if len(connStr) == 0 {
+				err = errors.New("cannot read the environment variable")
+			} else {
+				fmt.Println()
+				fmt.Println("DATABASE")
+				db = &dbPostgres{}
+				err = ensureDB(db, connStr)
+				if err == nil {
+
+					// Show the current state.
+					v, err := db.getCurrentVersion()
+					if err == nil {
+						fmt.Println("  Database is at migration version", v)
+					}
+				}
+			}
 		}
 	}
 
-	// Handle any errors bubbled up.
+	// Handle any errors.
 	if err != nil {
 		fmt.Println()
 		fmt.Println("ERROR")