+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+- 2024-03-04
+  - Connect to the database
+    - PostgreSQL only currently
+  - Create the `migratable_state` table
+  - Return the current migration version
+
 - 2024-03-03
   - Initial project creation
   - Standard repo files added